+++ impl/src/fmt/mod.rs
@@ -9,7 +9,7 @@ pub(crate) mod display;
 mod parsing;
 
 use proc_macro2::TokenStream;
-use quote::{format_ident, ToTokens};
+use quote::{format_ident, quote, ToTokens};
 use syn::{
     parse::{Parse, ParseStream},
     punctuated::Punctuated,
@@ -113,14 +113,16 @@ impl Parse for FmtAttribute {
     fn parse(input: ParseStream<'_>) -> syn::Result<Self> {
         Self::check_legacy_fmt(input)?;
 
-        Ok(Self {
+        let mut parsed = Self {
             lit: input.parse()?,
             comma: input
                 .peek(token::Comma)
                 .then(|| input.parse())
                 .transpose()?,
             args: input.parse_terminated(FmtArgument::parse, token::Comma)?,
-        })
+        };
+        parsed.args.pop_punct();
+        Ok(parsed)
     }
 }
 
@@ -271,6 +273,33 @@ impl FmtAttribute {
             .as_deref()
                 == Some(name)
         })
+    }
+
+    /// Returns an [`Iterator`] over the additional formatting arguments doing the dereferencing
+    /// replacement in this [`FmtAttribute`] for those [`Placeholder`] representing the provided
+    /// [`syn::Fields`] and requiring it
+    fn additional_deref_args<'fmt: 'ret, 'fields: 'ret, 'ret>(
+        &'fmt self,
+        fields: &'fields syn::Fields,
+    ) -> impl Iterator<Item = TokenStream> + 'ret {
+        let used_args = Placeholder::parse_fmt_string(&self.lit.value())
+            .into_iter()
+            .filter_map(|placeholder| match placeholder.arg {
+                Parameter::Named(name) => Some(name),
+                _ => None,
+            })
+            .collect::<Vec<_>>();
+
+        fields
+            .fmt_args_idents()
+            .into_iter()
+            .filter_map(move |field_name| {
+                (used_args.iter().any(|arg| field_name == arg)
+                    && !self.args.iter().any(|arg| {
+                        arg.alias.as_ref().map_or(false, |(n, _)| n == &field_name)
+                    }))
+                .then(|| quote! { #field_name = *#field_name })
+            })
     }
 
     /// Errors in case legacy syntax is encountered: `fmt = "...", (arg),*`.
@@ -523,6 +552,7 @@ where
     }
 }
 
+/// Extension of a [`syn::Type`] and a [`syn::Path`] allowing to travers its type parameters.
 trait ContainsGenericsExt {
     /// Checks whether this definition contains any of the provided `type_params`.
     fn contains_generics(&self, type_params: &[&syn::Ident]) -> bool;
@@ -635,6 +665,26 @@ impl ContainsGenericsExt for syn::Path {
                         }
                 }
             })
+    }
+}
+
+/// Extension of [`syn::Fields`] providing helpers for a [`FmtAttribute`].
+trait FieldsExt {
+    /// Returns an [`Iterator`] over [`syn::Ident`]s representing these [`syn::Fields`] in a
+    /// [`FmtAttribute`] as [`FmtArgument`]s or named [`Placeholder`]s.
+    ///
+    /// [`syn::Ident`]: struct@syn::Ident
+    // TODO: Return `impl Iterator<Item = syn::Ident> + '_` once MSRV is bumped up to 1.75 or
+    //       higher.
+    fn fmt_args_idents(&self) -> Vec<syn::Ident>;
+}
+
+impl FieldsExt for syn::Fields {
+    fn fmt_args_idents(&self) -> Vec<syn::Ident> {
+        self.iter()
+            .enumerate()
+            .map(|(i, f)| f.ident.clone().unwrap_or_else(|| format_ident!("_{i}")))
+            .collect()
     }
 }